+++ packages/validators/src/withMessages/alphaNum.js
@@ -1,5 +1,9 @@
 import alphaNum from '../raw/alphaNum'
 
+/**
+ * Validate if value is alpha-numeric string.
+ * @type {NormalizedValidator}
+ */
 export default {
   $validator: alphaNum,
   $message: 'The value must be alpha-numeric'