+++ packages/validators/src/withMessages/maxLength.js
@@ -1,15 +1,14 @@
 import maxLength from '../raw/maxLength'
-// import { withMessage, withParams } from '../common'
 
-export default (max) => ({
-  $validator: maxLength(max),
-  $message: ({ $params }) => `The maximum length allowed is ${$params.max}`,
-  $params: { max }
-})
-
-// Still figuring out which is less error prone
-
-// export default (max) => withMessage(
-//   withParams({ max }, maxLength),
-//   ({ $params }) => `The maximum length allowed is ${$params.max}`
-// )
+/**
+ * Validate the max length of a string.
+ * @param {Number} max
+ * @return {NormalizedValidator}
+ */
+export default function (max) {
+  return {
+    $validator: maxLength(max),
+    $message: ({ $params }) => `The maximum length allowed is ${$params.max}`,
+    $params: { max }
+  }
+}