+++ packages/validators/src/withMessages/minLength.js
@@ -1,7 +1,14 @@
 import minLength from '../raw/minLength'
 
-export default length => ({
-  $validator: minLength(length),
-  $message: ({ $params }) => `This field should be at least ${$params.length} long.`,
-  $params: { length }
-})
+/**
+ * Check if value is above a threshold.
+ * @param {Number | Ref<Number>} min
+ * @returns {NormalizedValidator}
+ */
+export default function (min) {
+  return {
+    $validator: minLength(min),
+    $message: ({ $params }) => `This field should be at least ${$params.min} long.`,
+    $params: { min }
+  }
+}