+++ packages/validators/src/withMessages/numeric.js
@@ -1,5 +1,9 @@
 import numeric from '../raw/numeric'
 
+/**
+ * Check whether a value is numeric.
+ * @type NormalizedValidator
+ */
 export default {
   $validator: numeric,
   $message: 'Value must be numeric'