+++ packages/vuelidate/test/unit/utils.js
@@ -1,4 +1,4 @@
-import { h, toRefs, isReactive } from 'vue'
+import { h } from 'vue'
 import { mount } from '@vue/test-utils'
 import { useVuelidate } from '../../src'
 
@@ -10,7 +10,7 @@ export const createSimpleComponent = (getVuelidateResults, state) => ({
     return {
       v,
       // spread the state so we have access to it
-      ...(isReactive(state) ? toRefs(state) : state)
+      state
     }
   },
   render () {
@@ -18,11 +18,13 @@ export const createSimpleComponent = (getVuelidateResults, state) => ({
   }
 })
 
-export const createSimpleWrapper = (rules, state) => mount(createSimpleComponent(() => useVuelidate(rules, state), state))
+export const createSimpleWrapper = (rules, state, config = {}) => mount(createSimpleComponent(() => useVuelidate(rules, state, config), state))
 
 export const shouldBePristineValidationObj = (v) => {
   expect(v).toHaveProperty('$error', false)
+  expect(v).toHaveProperty('$error', false)
   expect(v).toHaveProperty('$errors', [])
+  expect(v).toHaveProperty('$silentErrors', [])
   expect(v).toHaveProperty('$invalid', false)
   expect(v).toHaveProperty('$pending', false)
   expect(v).toHaveProperty('$dirty', false)
@@ -32,8 +34,35 @@ export const shouldBePristineValidationObj = (v) => {
 }
 
 export const shouldBeInvalidValidationObject = ({ v, property, propertyPath = property, validatorName }) => {
+  expect(v).toHaveProperty('$error', false)
+  expect(v).toHaveProperty('$errors', [])
+  expect(v).toHaveProperty('$silentErrors', [{
+    $message: '',
+    $params: {},
+    $pending: false,
+    $property: property,
+    $propertyPath: propertyPath,
+    $validator: validatorName
+  }])
+  expect(v).toHaveProperty('$invalid', true)
+  expect(v).toHaveProperty('$pending', false)
+  expect(v).toHaveProperty('$dirty', false)
+  expect(v).toHaveProperty('$anyDirty', false)
+  expect(v).toHaveProperty('$touch', expect.any(Function))
+  expect(v).toHaveProperty('$reset', expect.any(Function))
+}
+
+export const shouldBeErroredValidationObject = ({ v, property, propertyPath = property, validatorName }) => {
   expect(v).toHaveProperty('$error', true)
   expect(v).toHaveProperty('$errors', [{
+    $message: '',
+    $params: {},
+    $pending: false,
+    $property: property,
+    $propertyPath: propertyPath,
+    $validator: validatorName
+  }])
+  expect(v).toHaveProperty('$silentErrors', [{
     $message: '',
     $params: {},
     $pending: false,